Convert signals from strings to objects.
Gets rid of the signal-type/args split in function signatures, and the
whole "nested args" mess.  Signal handler functions no longer have to
parse `$args` based on the type of `$signal`: they simply receive an
object (whose class is now considered the signal type) with any
necessary arguments carried as properties of the class.

Signal handler definition uses the class names in the optional array of
signals to be handled, as well.

`Signal::send()` takes a `SignalInterface` instead of string (and
optional array).  The error and warning methods create new Error and
Warning objects, which implement SignalInterface.

A new method `Signal::sendError()` takes a pre-generated `Error` and
applies the standard "exit if no restart is invoked" logic, allowing
callers to easily use their own Error subclasses with that standord
process.